fix contollers according to frontend

+++ src/controllers/like.controller.js
@@ -24,12 +24,22 @@ const toggleVideoLike = asyncHandler(async (req, res) => {
         video: video._id,
         likedBy: req.user._id
     });
-
+    
     if (like) {
-        await Like.findByIdAndDelete(like._id);
+        if (like.isLike) {
+            await Like.findByIdAndDelete(like._id);
 
-        return res.status(200)
-        .json(new ApiResponse(200, "Video Like removed successfully"));
+            return res.status(200)
+            .json(new ApiResponse(200, "Video Like removed successfully")); 
+        } else {
+            const editedLike = await Like.findByIdAndUpdate(like._id, 
+                {
+                    isLike: true
+                }, { new: true }
+            )
+            return res.status(200)
+            .json(new ApiResponse(200, editedLike, "Video Like added successfully")); 
+        }
     } else {
         const createdLike = await Like.create({
             video: video._id,
@@ -38,6 +48,53 @@ const toggleVideoLike = asyncHandler(async (req, res) => {
 
         return res.status(200)
         .json(new ApiResponse(200, createdLike, "Video Like added successfully"));
+    }
+});
+
+const toggleVideoDislike = asyncHandler(async (req, res) => {
+    const { videoId } = req.params
+    //TODO: toggle like on video
+    if (!videoId) {
+        throw new ApiError(400, "Video ID is required");
+    }
+
+    const video = await Video.findById(videoId);
+
+    if (!video) {
+        throw new ApiError(404, "Video not found");
+    }
+
+    const like = await Like.findOne({
+        video: video._id,
+        likedBy: req.user._id
+    });
+
+
+    if (like) {
+        if (like?.isLike) {
+            const editedLike = await Like.findByIdAndUpdate(like._id, 
+                {
+                    isLike: false
+                }, { new: true }
+            )
+    
+            return res.status(200)
+            .json(new ApiResponse(200, editedLike, "Video Dislike added successfully"));
+        } else {
+            await Like.findByIdAndDelete(like._id);
+
+            return res.status(200)
+            .json(new ApiResponse(200, "Video Dislike removed successfully")); 
+        } 
+    } else {
+        const createdDislike = await Like.create({
+            video: video._id,
+            likedBy: req.user._id,
+            isLike: false
+        });
+
+        return res.status(200)
+        .json(new ApiResponse(200, createdDislike, "Video Dislike added successfully"));
     }
 });
 
@@ -116,7 +173,8 @@ const getLikedVideos = asyncHandler(async (req, res) => {
     const like = await Like.aggregate([
         {
             $match: {
-                likedBy: req.user._id
+                likedBy: req.user._id,
+                isLike: true
             }
         },
         {
@@ -168,14 +226,121 @@ const getLikedVideos = asyncHandler(async (req, res) => {
             }
         }
     ]);
-    
     return res.status(200)
-    .json(new ApiResponse(200 , like[0].liked_videos, "Liked Videos fetched successfully"));
+    .json(new ApiResponse(200 , like[0]?.liked_videos, "Liked Videos fetched successfully"));
+});
+
+const getDisikedVideos = asyncHandler(async (req, res) => {
+    //TODO: get all liked videos
+    const like = await Like.aggregate([
+        {
+            $match: {
+                likedBy: req.user._id,
+                isLike: false
+            }
+        },
+        {
+            $lookup: {
+                from: "videos",
+                localField: "video",
+                foreignField: "_id",
+                as: "liked_videos"
+            }
+        },
+        {
+            $unwind: "$liked_videos"
+        },
+        {
+            $lookup: {
+                from: "users",
+                localField: "liked_videos.owner",
+                foreignField: "_id",
+                as: "liked_videos.owner"
+            }
+        },
+        {
+            $addFields: {
+                "liked_videos.owner": { $arrayElemAt: ["$liked_videos.owner", 0] }
+            }
+        },
+        {
+            $group: {
+                _id: null,
+                liked_videos: { $push: "$liked_videos" }
+            }
+        },
+        {
+            $project: {
+                _id: 0,
+                liked_videos: {
+                    $map: {
+                        input: "$liked_videos",
+                        as: "video",
+                        in: {
+                            _id: "$$video._id",
+                            thumbnail: "$$video.thumbnail",
+                            title: "$$video.title",
+                            views: "$$video.views",
+                            owner: "$$video.owner.username"
+                        }
+                    }
+                }
+            }
+        }
+    ]);
+    return res.status(200)
+    .json(new ApiResponse(200 , like[0]?.liked_videos, "Liked Videos fetched successfully"));
+});
+
+const getVideoLike = asyncHandler(async (req, res) => {
+    const { videoId } = req.params;
+
+    if (!videoId) {
+        throw new ApiError(400, "Video ID is required");
+    }
+
+
+    const videoLike = await Like.aggregate([
+        {
+            $match: {
+                video: new mongoose.Types.ObjectId(videoId),
+                isLike: true
+            }
+        }
+    ]);
+
+    res.status(200)
+    .json(new ApiResponse(200, videoLike, "Video Like fetched successfully"));
+});
+
+const getVideoDislike = asyncHandler(async (req, res) => {
+    const { videoId } = req.params;
+
+    if (!videoId) {
+        throw new ApiError(400, "Video ID is required");
+    }
+
+
+    const videoLike = await Like.aggregate([
+        {
+            $match: {
+                video: new mongoose.Types.ObjectId(videoId),
+                isLike: false
+            }
+        }
+    ]);
+
+    res.status(200)
+    .json(new ApiResponse(200, videoLike, "Video Dislike fetched successfully"));
 });
 
 export {
     toggleCommentLike,
     toggleTweetLike,
     toggleVideoLike,
-    getLikedVideos
+    toggleVideoDislike,
+    getLikedVideos,
+    getDisikedVideos,
+    getVideoLike,
+    getVideoDislike
 }
